AbstractLoaderSessionInitiator implemented
